[수연][infinite-refactoring] Infinite scroll 개선 - hashtag 검색 시 깜빡임 제거, Api 호출 시간 단축
