+++ FE/package.json
@@ -39,7 +39,6 @@
     "react-quill": "^1.3.5",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
-    "react-scroll-horizontal": "^1.6.6",
     "simple-peer": "^9.11.1",
     "socket.io": "^4.5.1",
     "socket.io-client": "^4.5.1",
